Code Refactoring and Deduplication
